Add escape hatch for interpreting args literally
Trying to interpret all args as scripts is just a huge can of worms that
can make the whole parser fall over because you've got a braced string
with random stuff in it that's obviously not valid Tcl. Enter Tcl's
identity function: `string cat`. We can use it essentially "tag"
arguments that we want to receive special treatmeant (like being
interpreted literally).