Completely punt on parsing array indexes in bare words
Too hard at this point to have it work correctly when combined with
concatting ( or ) in bare words. It'd be nice to have it back at some
point, but will need to solve the issue of determining whether a
following ( should be concatted or not. This requires scanning to
the end of the word to see if it ends in ).